refactor: move enum classes to constant package
