+++ src/main/java/com/samyookgoo/palgoosam/payment/controller/request/PaymentConfirmRequest.java
@@ -1,6 +1,6 @@
 package com.samyookgoo.palgoosam.payment.controller.request;
 
-import com.samyookgoo.palgoosam.payment.domain.PaymentType;
+import com.samyookgoo.palgoosam.payment.constant.PaymentType;
 import lombok.Data;
 
 @Data
