+++ src/main/java/com/samyookgoo/palgoosam/payment/service/PaymentService.java
@@ -11,6 +11,7 @@ import com.samyookgoo.palgoosam.config.TossPaymentsConfig;
 import com.samyookgoo.palgoosam.deliveryaddress.domain.DeliveryAddress;
 import com.samyookgoo.palgoosam.deliveryaddress.dto.DeliveryAddressResponseDto;
 import com.samyookgoo.palgoosam.deliveryaddress.repository.DeliveryAddressRepository;
+import com.samyookgoo.palgoosam.payment.constant.PaymentStatus;
 import com.samyookgoo.palgoosam.payment.controller.request.PaymentConfirmRequest;
 import com.samyookgoo.palgoosam.payment.controller.request.PaymentCreateRequest;
 import com.samyookgoo.palgoosam.payment.controller.request.PaymentFailRequest;
@@ -18,7 +19,6 @@ import com.samyookgoo.palgoosam.payment.controller.response.OrderResponse;
 import com.samyookgoo.palgoosam.payment.controller.response.PaymentConfirmResponse;
 import com.samyookgoo.palgoosam.payment.controller.response.PaymentResponse;
 import com.samyookgoo.palgoosam.payment.domain.Payment;
-import com.samyookgoo.palgoosam.payment.domain.PaymentStatus;
 import com.samyookgoo.palgoosam.payment.repository.PaymentRepository;
 import com.samyookgoo.palgoosam.user.domain.User;
 import java.time.OffsetDateTime;
